Begin meta-definition of instructions
Many can be implemented in a generic method that is just told the
operand registers and values to work on based on info we can store about
each opcode.
The dispatch for instructions in the CPU will decode the opcode, look up
its entry in OPCODES, and then `send` to the value of the `method` key.

That's the plan, right now, anyway. There is a concern about lookup time
plus the overhead of method lookup, but that will need to be witnessed
in the real world before more decisions can be made.